feat: Gives the functionality of changing shelf to a book now reading, want to read, red.
The functiolnality of changing the shelf is made by changing the state of the component,
the book in this case. The book implemented to have state of the shelf for the apropriate shelf
and state none for out of the bookshelf. Change book shelf is the method of this component to use
in order to change shelf to a book or get a new book in bookshelf.
The book html code taken from the main app code to the component to render every book instance.